Removed Manager Class, Added IUpdateable interface, Moved Window Handling Function in own class.
